+++ notification-service/src/main/java/com/flight_booking/notification_service/application/service/EmailServiceImpl.java
@@ -2,13 +2,12 @@ package com.flight_booking.notification_service.application.service;
 
 import com.flight_booking.notification_service.domain.model.Notification;
 import com.flight_booking.notification_service.domain.repository.NotificationRepository;
+import com.flight_booking.notification_service.infrastructure.mail.MailSender;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
 
 @Slf4j
 @Service
