+++ notification-service/src/main/java/com/flight_booking/notification_service/infrastructure/mail/MailSender.java
@@ -1,4 +1,4 @@
-package com.flight_booking.notification_service.application.service;
+package com.flight_booking.notification_service.infrastructure.mail;
 
 // 이메일 전송용 인터페이스
 public interface MailSender {
